# backend/src/main/java/by/gusto/auth/dto/LoginResponse.java
package by.gusto.auth.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;
import lombok.experimental.SuperBuilder;

@Data
@EqualsAndHashCode(callSuper = true)
@NoArgsConstructor
@AllArgsConstructor
@SuperBuilder
public class LoginResponse extends TokenResponse {

    private UserResponse user;
}