# backend/src/main/java/by/gusto/auth/dto/PasswordResetRequest.java
package by.gusto.auth.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class PasswordResetRequest {

    @NotBlank
    @Email
    private String email;
}