# backend/src/main/java/by/gusto/auth/dto/TokenResponse.java
package by.gusto.auth.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.experimental.SuperBuilder;

@Data
@NoArgsConstructor
@AllArgsConstructor
@SuperBuilder
public class TokenResponse {

    private String accessToken;
    private String tokenType;
    private long expiresIn;
}